requirements.txt
python-telegram-bot
gunicorn==21.2.0
python-dotenv==1.0.1
